+++ src/components/index.ts
@@ -18,6 +18,10 @@ export { FFocusWarning } from './FFocusWarning.js';
 export { FDropOverlay, type DropFileSelectedDetail } from './FDropOverlay.js';
 export { FInfoDialog, type InfoDialogCloseDetail } from './FInfoDialog.js';
 export {
+  FConfirmDialog,
+  type ConfirmDialogResultDetail,
+} from './FConfirmDialog.js';
+export {
   FMenuPopup,
   type MenuActionDetail,
   type MenuActionName,